Verfiy that settings names are known
Issue #13 was due to the fact that the renaming of the `cycle_cpus`
setting to `pin_cpus` had not been updated in the configs. Therefore,
the configs still had a `cycle_cpus` key which was silently ignored, and
CPU cores were arbitrarily assigned by the system.

This wasn't the first time this mistake has happened, so in the future
all config names will be verified, and arbitrary keys can't be put into
config files.

Anyone adding new config options will need to add the new keys to the
list of strings in `src/parse_config.c`'s functions.