Speed up GlobalTimer64() by a few instructions
In order to record events as accurately as possible down to the
sub-microsecond scale, very fast clock sampling logic is necessary.

This commit shaves down the number of SASS instructions necessary
by four on sm_53 (Maxwell) and two on sm_70 (Volta), with savings
also found on other architectures (such as sm_62, Pascal).

Note that it is _not_ beneficial to rewrite the whole function in
PTX. Minimizing the number of PTX instructions was found to
increase the number of resultant SASS instructions, so this commit
leaves as much of the function as possible in C.

Further optimizations may be possible for sm_70+, as these
architectures can natively support 64-bit instructions, and do not
seem to have the same clock rollover bug.