Shrink `view_blocksbysm.py` block labels
GPUs now have far more SMs than when this script was originally
written, making the labels unreadably overlap in many situations.

This removes the space between the kernel name and the block number,
and shrinks the label font size to 10pt. 10pt is the same as the
IEEEtran footnote size, making it acceptably readable when inserted
in a paper (barring any intermediate scaling).