Add a new configuration option "terminator"
If set on a benchmark, the conclusion of that benchmark (due to
completing its iterations or exceeding its time limit), will
trigger all other benchmarks to exit once they complete their
current iteration.

This is useful when runing competing benchmarks against a benchmark
of interest. The competing benchmarks run asynchronously and
infinitely until the benchmark of interest completes, then they are
terminated.